Cast int

+++ logic/chain_iterator.py
@@ -13,9 +13,9 @@ class ChainIterator:
     def process_contacts(contact_group):
         for contact in contact_group:
             if contact[1] > contact[3]:
-                Database.update_risk_level(contact[2], (contact[3] + contact[1] * contact[5]) % 5)
+                Database.update_risk_level(contact[2], int(contact[3] + contact[1] * contact[5]) % 5)
             else:
-                Database.update_risk_level(contact[0], (contact[1] + contact[3] * contact[5]) % 5)
+                Database.update_risk_level(contact[0], int(contact[1] + contact[3] * contact[5]) % 5)
 
     # Process new contacts with level 5s of the last hour
     @staticmethod
